requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
pydantic
pydantic-settings
passlib[bcrypt]
python-jose[cryptography]
stripe
celery
redis
requests